Minor types in README.md
Just fixing a few typos.

Also, was thinking that it might be worth mentioning earlier in the README how the script works (that the LLM generates conversion code). I know you mention that in the “Security” section at the bottom of the README, but I just thought it was cool and would be nice to see that discussed earlier.